- Add `AccThunk` type (as suggested in https://github.com/FluxML/Zygote.jl/pull/1555) that holds thunks to-be-accumulated without unthunking prematurely.
- Add reduced MWE from https://github.com/FluxML/Flux.jl/issues/2585 as test, that leads to infinite compilation without `AccThunk`.

### PR Checklist

- [x] Tests are added