@pxl-th 2025-03-13
> Coming from [#1555 (comment)](https://github.com/FluxML/Zygote.jl/pull/1555#issuecomment-2611733494), I think we need at least one test that shows we aren't introducing new type instabilities when multiple paths try to accumulate different (un)thunked types.

IIUC you mean the case when `AccThunk` holds thunks of different type and if that leads to more type-instability -> worse performance?