@ToucheSir 2025-03-17
I was thinking something which involves smaller input sizes. e.g. this regression-esque example:
```julia
using Zygote, LinearAlgebra

x = rand(10, 100)
w = rand(2, 10)

f(w, x) = sum(w * x) + norm(w)

@benchmark gradient(f, $w, $x)
```