@pxl-th 2025-03-17
On your example this PR is slower:

- This PR:
```julia
BenchmarkTools.Trial: 10000 samples with 6 evaluations per sample.
 Range (min … max):  5.714 μs …  1.974 ms  ┊ GC (min … max):  0.00% … 99.07%
 Time  (median):     8.062 μs              ┊ GC (median):     0.00%
 Time  (mean ± σ):   9.420 μs ± 39.797 μs  ┊ GC (mean ± σ):  13.65% ±  3.90%
     ▂██▇▃             ▃▂                                     
  ▁▂▅█████▇▅▄▃▃▂▂▃▂▄▅▆█████▇▆▄▃▃▂▂▂▁▁▁▁▁▁▁▂▂▂▂▃▄▃▃▄▃▃▃▂▂▂▂▂▁ ▃
  5.71 μs        Histogram: frequency by time        12.8 μs <
 Memory estimate: 13.47 KiB, allocs estimate: 47.
```
- `master`:
```julia
BenchmarkTools.Trial: 10000 samples with 8 evaluations per sample.
 Range (min … max):  3.047 μs … 603.206 μs  ┊ GC (min … max):  0.00% … 98.22%
 Time  (median):     3.696 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   4.684 μs ±  11.762 μs  ┊ GC (mean ± σ):  10.44% ±  5.14%
   ▃▆▇▇██▇▆▄▃▂▂▁▁                   ▁▁▁▁▁▁▁▁▁         ▁▁▁▂▁   ▂
  ▆█████████████████▇▆█▇▆▆▅▆▅▅▃▃▂▃▅▇█████████▇▇▆▆▆▇▇▇████████ █
  3.05 μs      Histogram: log(frequency) by time      9.14 μs <
 Memory estimate: 11.91 KiB, allocs estimate: 15.
```

But the times are so small it is not obvious if it's actually 2x slower on other examples.

Trying out reduced MWE that I've added as a test:

```julia
function main()
    W = ones(Float32, 10, 10)
    x = [ones(Float32, 10) for i in 1:512]
    gs = gradient(W) do W
        sum((W * xi)[1] for xi in x)
    end
    return
end
@benchmark main()
```

- This PR:
```julia
BenchmarkTools.Trial: 438 samples with 1 evaluation per sample.
 Range (min … max):   9.155 ms … 21.202 ms  ┊ GC (min … max):  0.00% … 11.38%
 Time  (median):     11.008 ms              ┊ GC (median):    14.24%
 Time  (mean ± σ):   11.387 ms ±  1.351 ms  ┊ GC (mean ± σ):  13.36% ±  3.84%
           ▅█▅▇▄                                               
  ▃▃▂▂▃▄▄▅▆██████▅▆▅▅▆▄▄▄▄▃▂▅▁▃▂▂▃▃▂▁▃▁▃▁▂▁▁▂▂▁▁▁▁▂▁▁▁▁▁▁▁▁▁▃ ▃
  9.15 ms         Histogram: frequency by time        17.9 ms <
 Memory estimate: 31.41 MiB, allocs estimate: 49627.
```
- `master`:
```julia
BenchmarkTools.Trial: 395 samples with 1 evaluation per sample.
 Range (min … max):  10.143 ms … 23.214 ms  ┊ GC (min … max):  0.00% … 29.74%
 Time  (median):     12.530 ms              ┊ GC (median):     0.00%
 Time  (mean ± σ):   12.623 ms ±  2.049 ms  ┊ GC (mean ± σ):  10.17% ± 10.01%
    ▄█  ▁                                                      
  ▃████▅█▅▆▆▅▅▄▃▃▃▃▃▃▃▅▅▇▆▅▆▆█▅▃▅▃▅▄▅▅▄▅▃▄▃▃▃▃▃▃▄▂▃▃▁▂▁▂▂▁▂▃▃ ▃
  10.1 ms         Histogram: frequency by time        17.6 ms <
 Memory estimate: 11.82 MiB, allocs estimate: 64480.
```