@ToucheSir 2025-03-18
My main worry would be that this introduces allocations into workloads which were low to zero allocation before. The loop example is perhaps the best case scenario for this change, because it involves a large number of accumulated thunks which all have the same type.

My other worry is that this makes debugging Zygote code quite a bit harder. Currently Cthulhu.jl works well because a lot of functions can be statically inferred. But with this change, many functions which were type stable under differentiation will no longer be, so that tool will not be available. The change probably makes sense on balance, but losing access to such an important debugging tool (really the _only_ debugging tool, because the Debugger.jl does not work well enough) makes me uncomfortable.

Anyhow, I'll defer to @mcabbott or @CarloLucibello for a second opinion on this.