@pxl-th 2025-03-18
> The change probably makes sense on balance, but losing access to such an important debugging tool (really the only debugging tool, because the Debugger.jl does not work well enough) makes me uncomfortable.

We can make this opt-in then and disable thunks by default.